Sharing ProjectSettingsFactory between different versions of ReSharper.
