+++ Project/Src/AddIns/ReSharper611/ReSharper611.csproj
@@ -234,7 +234,6 @@
     <Compile Include="CodeCleanup\Descriptors\DocumentationDescriptor.cs" />
     <Compile Include="Options\HighlightingRegistering.cs" />
     <Compile Include="ShellComponents\StyleCopCodeStyleChecker.cs" />
-    <Compile Include="Core\ProjectSettingsFactory.cs" />
     <Compile Include="Core\DeclarationHeader.cs" />
     <Compile Include="Core\FileHeader.cs" />
     <Compile Include="Core\StyleCopCoreFactory.cs" />
