+++ Project/Src/AddIns/ReSharper710/Core/StyleCopCoreFactory.cs
@@ -19,8 +19,9 @@ namespace StyleCop.ReSharper710.Core
 {
     #region Using Directives
 
+    using ReSharperBase.Core;
+
     using StyleCop.Diagnostics;
-    using SourceCodeFactory = ReSharperBase.Core.SourceCodeFactory;
 
     #endregion
